+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to the
 
 ## [0.3.0.0] - 
 
+### Fixed
+
+- `getDocList`: When the list of requested fields was not exhaustive
+  to parse into the specified record type, `getDocList` ended up to
+  parse the response into `Ok _ _ []` instead of `Err _ _`. With this
+  bug the response looked like we got no results but in reality it was
+  just wrong API use.
+
 ### Changed
 
 - Fix names in public API because these functions don't operate on
@@ -18,8 +26,15 @@ and this project adheres to the
 
 ### Added
 
+- `getAllFields` to fetch names of all fields for a given DocType.
 - `getDocListAllFields` to fetch list of documents with all their
   fields without explicitly naming them.
+- `systemFieldnames` to define some fixed fieldnames like `_user_tags` and `_assign`.
+- `IsString` instance for `FilterValue` so that `"my value"` is
+  automatically converted to `FilterText "my value"`.
+- `showJsonPretty` and `showJsonResponsePretty` for debugging and
+  pretty-printing JSON API responses (adds dependency `aeson-pretty`.
+- `andThen` to concatenate two API calls passing the result of A directly to B.
 
 ## [0.2.1.0] - 2026-03-13